[update] : Chipコンポーネントのディレクトリを移動
